Renamed scene builder.
