+++ bindings/wasm/lib/scene-builder.ts
