+++ bindings/wasm/package.json
@@ -24,8 +24,8 @@
     "lib/debug.js",
     "lib/evaluate.d.ts",
     "lib/evaluate.js",
-    "lib/export.d.ts",
-    "lib/export.js",
+    "lib/scene-builder.d.ts",
+    "lib/scene-builder.js",
     "lib/gltf-io.d.ts",
     "lib/gltf-io.js",
     "lib/3mf-export.d.ts",
